requirements.txt
asgiref==3.8.1
chardet==5.2.0
dj-database-url==2.1.0
Django==5.1.5
django-model-utils==5.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==22.0.0
mysqlclient==2.2.6
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
reportlab==4.3.0
sqlparse==0.5.3
typing_extensions==4.12.2
wheel==0.38.1
whitenoise==6.6.0
